requirements.txt
absl-py==1.1.0
anyio==3.6.1
asgiref==3.5.2
astunparse==1.6.3
black==22.3.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.78.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h11==0.13.0
h5py==3.7.0
idna==3.3
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pyparsing==3.0.9
python-multipart==0.0.5
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
Werkzeug==2.1.2
wrapt==1.14.1
